// projects/step-core/src/lib/client/generated/models/Event.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Event = {
    id?: string;
    name?: string;
    group?: string;
    payload?: any;
    creationTimestamp?: number;
    submitionTimestamp?: number;
    receptionTimestamp?: number;
    insertionTimestamp?: number;
    deletionTimestamp?: number;
    lastReadTimestamp?: number;
};